@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.5.10'
+    ext.kotlin_version = '1.6.10'
 
     repositories {
         mavenCentral()
@@ -19,7 +19,7 @@ buildscript {
 }
 
 group 'net.skripthub'
-version '1.4'
+version '1.5'
 
 apply plugin: 'com.github.johnrengelman.shadow'
 apply plugin: 'java'
@@ -38,11 +38,12 @@ repositories {
     maven { url = 'https://oss.sonatype.org/content/repositories/central' }
     maven { url 'https://repo.destroystokyo.com/repository/maven-public/' }
     maven { url 'https://papermc.io/repo/repository/maven-public/' }
+    maven { url 'https://repo.skriptlang.org/releases' }
     maven { url 'https://jitpack.io' }
 }
 
 dependencies {
-    implementation 'com.github.SkriptLang:Skript:2.6'
+    implementation 'com.github.SkriptLang:Skript:master-SNAPSHOT'
 
     implementation group: 'org.spigotmc', name: 'spigot-api', version:'1.17-R0.1-SNAPSHOT'
 
