@@ -7,7 +7,8 @@ data class AddonData(var name: String,
                      var expressions: MutableList<SyntaxData> = mutableListOf(),
                      var types: MutableList<SyntaxData> = mutableListOf(),
                      var functions: MutableList<SyntaxData> = mutableListOf(),
-                     var sections: MutableList<SyntaxData> = mutableListOf()) {
+                     var sections: MutableList<SyntaxData> = mutableListOf(),
+                     var structures: MutableList<SyntaxData> = mutableListOf()) {
     fun sortLists() {
         events.sortBy { info -> info.name }
         conditions.sortBy { info -> info.name }
@@ -16,5 +17,6 @@ data class AddonData(var name: String,
         types.sortBy { info -> info.name }
         functions.sortBy { info -> info.name }
         sections.sortBy { info -> info.name }
+        structures.sortBy { info -> info.name }
     }
 }
